+++ RELEASES.md
@@ -2,6 +2,12 @@
 
 ## 1.0.7
 
+* fontconv now notifies users of requested characters unavailable in the font.
+* Renamed --points parameter to --pixels.
+* Improved documentation.
+
+## 1.0.7
+
 * Ported fontconv tool to Python.
 * Added textproc tool for calculating Unicode codepoint sets from text/.c files.
 * Fixed examples so they compile correctly.